requirement.txt
amqp==1.4.6
anyjson==0.3.3
billiard==3.3.0.20
celery==3.1.18
Flask==0.10.1
Flask-SocketIO==0.6.0
gevent==1.0.2
gevent-socketio==0.3.6
gevent-websocket==0.9.5
greenlet==0.4.7
itsdangerous==0.24
Jinja2==2.8
kombu==3.0.26
MarkupSafe==0.23
pytz==2015.4
PyYAML==3.11
redis==2.10.3
requests==2.7.0
Werkzeug==0.10.4
wheel==0.24.0
dnspython==1.12.0
peewee==2.6.3
psycopg2==2.6.1
arnold==0.4.0
pycountry==1.13